requirements.txt
torch>=1.12.0  # PyTorch version with MPS support (make sure this aligns with the latest MPS-compatible version)
torchvision>=0.13.0
transformers==4.25.1  # Hugging Face transformers library with Vision Transformer model
scikit-learn>=0.24  # For train-test splitting
Pillow>=9.0  # For handling images
requests>=2.25  # For downloading the dataset if needed
kaggle>=1.5.12  # Kaggle API for dataset download
kagglehub==0.3.4
pandas==2.2.3
matplotlib==3.9.2
opencv-python==4.10.0.84
seaborn==0.13.2
pyarrow==18.0.0
tensorflow==2.18.0
Keras-Preprocessing==1.1.2
safetensors==0.4.5
joblib==1.4.2
openai==1.54.3
geopy==2.4.1
python-dotenv==1.0.1
